requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
websockets>=12.0
httpx>=0.25
python-dotenv>=1.0
anthropic>=0.39
hidapi>=0.14
pillow>=10.0
